+++ sshd-common/src/main/java/org/apache/sshd/common/kex/KexState.java
@@ -24,15 +24,32 @@ import java.util.EnumSet;
 import java.util.Set;
 
 /**
- * Used to track the key-exchange (KEX) protocol progression
+ * Used to track the key-exchange (KEX) protocol progression.
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
 public enum KexState {
     UNKNOWN,
+
+    /**
+     * A key exchange has been initiated by this party.
+     */
     INIT,
+
+    /**
+     * A key exchange is running. Both sides have received the peer's SSH_MSG_KEX_INIT.
+     */
     RUN,
+
+    /**
+     * Keys have been exchanged; this party has sent its SSH_MSG_NEW_KEYS.
+     */
     KEYS,
+
+    /**
+     * The key exchange is done; both parties have received the peer's SSH:MSG_NEW_KEYS, and there is no ongoing key
+     * exchange.
+     */
     DONE;
 
     public static final Set<KexState> VALUES = Collections.unmodifiableSet(EnumSet.allOf(KexState.class));